# pythoncode/Ecommerce/Model/product.py
#!/usr/bin/python
# -*- coding: utf-8 -*-

class Product:
    def __init__(self):
        self.id = None
        self.name = None
        self.description = None
        self.price = None
        self.stockQuantity = None

    def isAvailable(self, ):
        pass

    def updateStock(self, quantity):
        pass